+++ src/apps/obsidian-plugin/PluginSettings.ts
@@ -3,6 +3,7 @@ import { ScheduledItemsSectionSelection } from "./views";
 
 export type SimpleBudgetHelperSettings = {
 	dbId: string;
+	defaultCurrency: string;
 	rootFolder: string;
 	debugMode: boolean;
 	lastTab: {
@@ -13,6 +14,7 @@ export type SimpleBudgetHelperSettings = {
 
 export const DEFAULT_SETTINGS: SimpleBudgetHelperSettings = {
 	dbId: "",
+	defaultCurrency: "USD",
 	rootFolder: "Budget",
 	debugMode: false,
 	lastTab: {